remove async trait from file loader
