+++ ikari/Cargo.toml
@@ -90,7 +90,6 @@ futures-intrusive = "0.5.0"
 rmp-serde = "1.1.2"
 serde = "1.0.173"
 zstd = { version = "0.12.4", default-features = false }
-async-trait = "0.1.73"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 basis-universal = "0.3.0"
