+++ ikari/src/lib.rs
@@ -12,6 +12,10 @@ pub mod camera;
 pub mod character;
 pub mod collisions;
 pub mod file_loader;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod file_loader_native;
+#[cfg(target_arch = "wasm32")]
+pub mod file_loader_wasm;
 pub mod game;
 pub mod game_state;
 pub mod gameloop;
